requirement.txt
pytest==4.4.0
pytest-mock==1.10.3
Click==7.0
kafka-python==1.4.6